TsBlog-6

+++ 依赖注入/TsBlog/src/Libraries/TsBlog.Services/PostAppServices/IPostAppService.cs
@@ -26,7 +26,7 @@ namespace TsBlog.Services.PostAppServices
         /// </summary>
         /// <param name="entity">博文实体类</param>
         /// <returns></returns>
-        int Insert(Post entity);
+        long Insert(Post entity);
 
         /// <summary>
         /// 更新实体数据
